Add Permission policy for Permission model

+++ app/Http/Controllers/PermissionController.php
@@ -8,19 +8,18 @@ use Illuminate\Http\Request;
 class PermissionController extends Controller
 {
 
-    //index permissions
-    /**
-     * PermissionController constructor.
-     */
-    public function __construct()
-    {
-        $this->authorizeResource(Permission::class, 'permission');
-    }
+//    //index permissions
+//    /**
+//     * PermissionController constructor.
+//     */
+//    public function __construct()
+//    {
+//        $this->authorizeResource(Permission::class, 'permission');
+//    }
 
     public function index()
     {
-        dd("index");
-        $this->authorize('viewAny', auth()->user());
+        $this->authorize('viewAny', Permission::class);
         $permissions = Permission::all();
         return view('permissions.index', compact('permissions'));
     }
@@ -28,7 +27,6 @@ class PermissionController extends Controller
     //form to create permissions
     public function create()
     {
-        dd("create");
         $this->authorize('create', Permission::class);
         return view('permissions.create');
     }
@@ -36,7 +34,6 @@ class PermissionController extends Controller
     //store permissions
     public function store()
     {
-        dd("store");
         $this->authorize('create', Permission::class);
         Permission::create(request()->validate([
             'name' => 'required',
@@ -47,21 +44,21 @@ class PermissionController extends Controller
     //show a single permission
     public function show(Permission $permission)
     {
-        $this->authorize('view', auth()->user(), $permission);
+        $this->authorize('view', $permission);
         return view('permissions.show', compact('permission'));
     }
 
     //edit form
     public function edit(Permission $permission)
     {
-        $this->authorize('update', auth()->user(), $permission);
+        $this->authorize('update', $permission);
         return view('permissions.edit', compact('permission'));
     }
 
     //update permission
     public function update(Permission $permission)
     {
-        dd("update");
+        $this->authorize('update', $permission);
         $data = request()->validate([
            'name' => 'required'
         ]);
@@ -71,7 +68,7 @@ class PermissionController extends Controller
     //delete a permission
     public function destroy(Permission $permission)
     {
-        dd("delete");
+        $this->authorize('delete', $permission);
         $permission->delete();
     }
 }
